orchestration/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0
anthropic>=0.40.0
httpx>=0.27.0
numpy>=1.26.0
aiosqlite>=0.20.0
dependency-injector>=4.41.0
alembic>=1.13.0
sqlalchemy>=2.0
PyJWT>=2.8.0
bcrypt>=4.0.0
slowapi>=0.1.9
email-validator>=2.0
authlib>=1.3.0
mcp>=1.0.0
